[#16] feat : api 모듈화
